% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_agegr1n}
\alias{format_agegr1n}
\title{Derive pooled age group (N)}
\usage{
format_agegr1n(x)
}
\description{
\code{format_agegr1n} derives the \code{AGEGR1N} variable, which is the numeric code for \code{AGEGR1}.
}
