% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_bmiblgr1}
\alias{format_bmiblgr1}
\title{Derive BMIBLGR1}
\usage{
format_bmiblgr1(x)
}
\description{
\code{format_bmiblgr1} derives the BMIBLGR1 variable, which pools subjects into baseline BMI groups.
}
