% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_dcsreas}
\alias{format_dcsreas}
\title{Derive DCSREAS}
\usage{
format_dcsreas(dsdecod)
}
\description{
\code{format_dcsreas} derives the DCSREAS variable, which contains subjects' disposition reason(s).
}
