% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_durdsgr1}
\alias{format_durdsgr1}
\title{Derive DURDSGR1}
\usage{
format_durdsgr1(x)
}
\description{
\code{format_durdsgr1} derives the DURDSGR1 variable, which pools subjects into disease duration groups.
}
