% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_functions.R
\name{format_dcsreas}
\alias{format_dcsreas}
\title{Derive subject's disposition reason(s)}
\usage{
format_dcsreas(dsdecod)
}
\description{
\code{format_dcsreas} derives the \code{DCSREAS} variable, which contains subjects' disposition reason(s).
}
